refactor(refactor): refactor Labeltag component
