--- client/src/components/Core/Text/LabelWithIcon.module.scss
@@ -1,10 +1,0 @@
-@import 'styles/variables/colors';
-@import 'styles/variables/spacing';
-
-.container {
-    display: flex;
-    align-items: center;
-    .icon {
-        margin-right: spacing(1);
-    }
-}
