+++ client/src/views/Authorized/Participants/mocks/participants.ts
@@ -11,7 +11,7 @@ import { LevelInformationFieldsetModel } from 'components/fieldsets/participants
 import { MotivationInformationFieldsetPrefillData } from 'components/fieldsets/participants/fieldsets/MotivationInformationFieldset'
 import { PermissionsFieldsetPrefillData } from 'components/fieldsets/participants/fieldsets/PermissionsFieldset'
 import { ReadingTestInformationFieldsetModel } from 'components/fieldsets/participants/fieldsets/ReadingTestInformationFieldset'
-import { RefererInformationFieldsetModel } from 'components/fieldsets/participants/fieldsets/ReferrerInformationFieldset'
+import { RefererInformationFieldsetModel } from 'components/fieldsets/participants/fieldsets/RefererInformationFieldset'
 import { WorkInformationFieldsetModel } from 'components/fieldsets/participants/fieldsets/WorkInformationFieldset'
 import { WritingInformationFieldsetModel } from 'components/fieldsets/participants/fieldsets/WritingInformationFieldset'
 import { PersonInformationFieldsetModel } from 'components/fieldsets/shared/PersonInformationFieldset'
