+++ client/src/views/Authorized/Participants/taalhuis/Participants/Detail/Intake/ParticipantReadView.tsx
@@ -31,7 +31,7 @@ import { ReadDetailTabs, readDetailTabsTranslations } from '../../../constants'
 import EducationInformationFieldset from '../../../../../../../components/fieldsets/participants/fieldsets/EducationInformationFieldset'
 import LevelInformationFieldset from '../../../../../../../components/fieldsets/participants/fieldsets/LevelInformationFieldset'
 import CivicIntegrationFieldset from '../../../../../../../components/fieldsets/participants/fieldsets/CivicIntegrationInformationFieldset'
-import RefererInformationFieldset from '../../../../../../../components/fieldsets/participants/fieldsets/ReferrerInformationFieldset'
+import RefererInformationFieldset from '../../../../../../../components/fieldsets/participants/fieldsets/RefererInformationFieldset'
 import WorkInformationFieldset from '../../../../../../../components/fieldsets/participants/fieldsets/WorkInformationFieldset'
 import MotivationInformationFieldset from '../../../../../../../components/fieldsets/participants/fieldsets/MotivationInformationFieldset'
 import ReadingTestInformationFieldset from '../../../../../../../components/fieldsets/participants/fieldsets/ReadingTestInformationFieldset'
